push: Forget about activated callback being an EventEmitter event.
A callback is fine. Even if the event is called from the service worker,
if that's even possible, it doesn't make a difference if we require the
user to register for an event vs. we require to call `activate` with a
fresh callback.

Also, I really meant that for useCustomRegisterer, not this callback.
But the same applies to useCustomRegisterer, really.